requirements.txt
flask
openai
python-dotenv
chromadb
#torch==2.1.0
torch==2.2.2 --index-url https://download.pytorch.org/whl/cpu
sentence-transformers
pdfplumber
python-docx
pandas
psycopg2-binary
python-dotenv
numpy==1.26.4
scikit-learn